+++ packages/stacks/firebase/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @elbwalker/stack-firebase
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [0c4d649]
+  - @elbwalker/utils@3.1.2
+  - @elbwalker/client-node@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes